Route protégée pour les administateurs

+++ routes/web.php
@@ -19,12 +19,12 @@ Route::get('/', 'HomeController@index')->name('home');
 
 Route::get('/actualites/create', 'ActualiteController@create')->name('actualiteCreate');
 Route::post('/actualites/create', 'ActualiteController@store')->name('actualiteStore');
-Route::get('/actualites/edit/{id}', 'ActualiteController@edit')->where(['id' => '[0-9]+'])->name('actualiteEdit');
-Route::post('/actualites/edit/{id}', 'ActualiteController@update')->where(['id' => '[0-9]+'])->name('actualiteUpdate');
-Route::delete('/actualites/{id}', 'ActualiteController@delete')->where(['id' => '[0-9]+'])->name('actualiteDelete');
+Route::get('/actualites/edit/{id}', 'ActualiteController@edit')->where(['id' => '[0-9]+'])->name('actualiteEdit')->middleware('admin');
+Route::post('/actualites/edit/{id}', 'ActualiteController@update')->where(['id' => '[0-9]+'])->name('actualiteUpdate')->middleware('admin');
+Route::delete('/actualites/{id}', 'ActualiteController@delete')->where(['id' => '[0-9]+'])->name('actualiteDelete')->middleware('admin');
 
 Route::post('/commentaires/create', 'CommentaireController@store')->name('commentaireStore');
-Route::get('/commentaires/delete/{id}', 'CommentaireController@delete')->where(['id' => '[0-9]+'])->name('commentaireDelete');
+Route::get('/commentaires/delete/{id}', 'CommentaireController@delete')->where(['id' => '[0-9]+'])->name('commentaireDelete')->middleware('admin');
 
 Route::get('/profil', 'ProfilController@index')->name('profil');
 Route::get('/profil/email', 'ProfilController@email')->name('profilEmail');
@@ -35,15 +35,15 @@ Route::post('/profil/mdp', 'ProfilController@changeMdp')->name('profilChangeMdp'
 Route::get('/evenements', 'EvenementController@index')->name('evenements');
 Route::get('/evenements/create', 'EvenementController@create')->name('evenementCreate');
 Route::post('/evenements/create', 'EvenementController@store')->name('evenementStore');
-Route::get('/evenements/edit/{id}', 'EvenementController@edit')->where(['id' => '[0-9]+'])->name('evenementEdit');
-Route::post('/evenements/edit/{id}', 'EvenementController@update')->where(['id' => '[0-9]+'])->name('evenementUpdate');
-Route::get('/evenements/delete/{id}', 'EvenementController@delete')->where(['id' => '[0-9]+'])->name('evenementDelete');
-//Route::get('/evenements/participer/{idEvenement}/{idParticipation}', 'EvenementController@participer')->name('evenementParticiper');
+Route::get('/evenements/edit/{id}', 'EvenementController@edit')->where(['id' => '[0-9]+'])->name('evenementEdit')->middleware('admin');
+Route::post('/evenements/edit/{id}', 'EvenementController@update')->where(['id' => '[0-9]+'])->name('evenementUpdate')->middleware('admin');
+Route::get('/evenements/delete/{id}', 'EvenementController@delete')->where(['id' => '[0-9]+'])->name('evenementDelete')->middleware('admin');
+
 Route::get('/evenements/participer/{idEvenement}/{idParticipation}', 'EvenementController@participer')->name('evenementParticiper');
 
 Route::get('/publications', 'PublicationController@index')->name('publication');
 Route::post('/publications/create', 'PublicationController@store')->name('publicationStore');
-Route::get('/publications/delete/{id}', 'PublicationController@delete')->where(['id' => '[0-9]+'])->name('publicationDelete');
+Route::get('/publications/delete/{id}', 'PublicationController@delete')->where(['id' => '[0-9]+'])->name('publicationDelete')->middleware('admin');
 
 Route::get('/profil/{id}', 'ProfilController@show')->where(['id' => '[0-9]+'])->name('profilShow');
